refactor and reorganize svelte-scoped transformSFC into multiple parts
